Add alternative shortcut for forward/rewind
This adds a new shortcut to rewind/forward a video, via alt + left/right
click, respectively. The increment is set to 0.1s by default, and can be
configured to up to 10s via the video controls screen.
The existing shortcut for rewinding/forwarding a video by 5s is left
unchanged.